Перенес логику из контроллера для участников в сервис
